### 2025-04-06 #8012 @editorialbot
[PRE REVIEW]: fgwqsr: An R package for Frequentist Grouped Weighted Quantile Sum Regression
**Submitting author:** @Daniel-Rud (<a href="http://orcid.org/0000-0002-0508-4552">Daniel Rud</a>)
**Repository:**  https://github.com/Daniel-Rud/fgwqsr
**Branch with paper.md** (empty if default branch): joss-branch
**Version:** v1.0.0
