### 2025-04-06 #8012 @Daniel-Rud
@jedbrown @arfon @xuanxu  I need to change the paper repository to https://github.com/Daniel-Rud/fgwqsr-JOSS .  Can this be done?  I thought when it had asked initially for a repository, it was asking for the software repository, my apologies
