### 2025-04-22 #8012 @Daniel-Rud
@nanhung Hi Nan-Hung, I would like to ask if you would be interested in reviewing our JOSS submission "fgwqsr: An R Package for Frequentist Grouped Weighted Quantile Sum Regression". We offer an implementation of the weighted quantile sum regression that does not require data splitting, has inference for both group effects and weights, and improves power to detect effects in chemical mixture settings. If interested, our mansucript has been published in Statistics in Medicine, https://onlinelibrary.wiley.com/doi/full/10.1002/sim.70078. Thank you very much for your consideration!
