### 2025-04-22 #8012 @danielskatz
👋 @Daniel-Rud - While I'm not the editor or track editor here, I do want to point out inviting reviewers is the job of the editor, once one is assigned. Your suggestions of potential reviewers are likely welcome. But we specifically would ask that you not tag potential reviewers, as they don't need to be notified and added to a thread until the editor chooses to invite them. See the first comment in this issue above, which includes

> @Daniel-Rud if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
