+++ README.Rmd
@@ -1,464 +1,0 @@
----
-output: github_document
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.path = "man/figures/README-",
-  out.width = "100%"
-)
-```
-
-# fgwqsr -- Frequentist Grouped Weighted Quantile Sum Regression
-Fully frequentist model fitting procedure for the weighted quantile sum regression formulation.  Included in this package is an implementation of Bayesian Grouped Weighted Quantile Sum Regression (bgwqsr) introduced by Wheeler, David C et al. that fits Markov Chain Monte Carlo (MCMC) chains in parallel through the runjags package.  
-
-<!-- badges: start -->
-<!-- badges: end -->
-
-
-# Overview 
-<br>
-FGWQSR estimates parameters to the following model: 
-
-<br>
-
-$$y_i \sim \text{Bernoulli}(\pi_i) $$
-$$\text{logit}(\pi_i) = c_0+ \sum\limits_{g = 1}^G \gamma_g \bigg( \sum\limits_{k = 1} ^{c_g} w_{g,k} \cdot q_{g,k,i}\bigg) + \sum_{r = 1}^R \phi_rz_{r,i} $$
-where 
-
-* $y_i$ - outcome variable (coded 0 1)
-
-* $c_0$ - offset variable (intercept)
-
-* $G$ - total number of chemical mixture groups 
-
-* $\gamma_g$ - group index effect corresponding to mixture group $g$
-
-* $c_g$ - total number of chemicals in group $g$
-
-* $w_{g,k}$ - weight for chemical $k$ in group $g$
-
-* $q_{g,k,i}$ - quantized chemical exposure $k$ in group $g$ for individual $i$ 
-
-* $R$ - total number of adjusting covariates 
-
-* $\phi_r$ - effect of $r^{\text{th}}$ adjusting covariate 
-
-* $z_{r,i}$ - measure of adjusting covariate $r$ for individual $i$
-
-
-subject to the constraints $$\sum_{k = 1}^{c_g} w_{g,k} =1, \ w_{g,k} \in (0,1)$$ without requiring any data splitting
-
-<br>
-
-
-
-# Installation
-
-You can install the development version of fgwqsr from [GitHub](https://github.com/) with:
-
-```{r}
-# install.packages("devtools")
-devtools::install_github("Daniel-Rud/fgwqsr")
-```
-
-```{r example}
-library(fgwqsr)
-```
-
-In order to use fgwqsr, let us first generate some data
-
-# Data Generation 
-<br>
-To begin, let us generate some data to use.  We will have 14 underlying chemical exposure variables, with mixture groups of sizes (5,4,5) and weight distribution   
-
-* $\mathbf{w}_1 = (1/3,1/3,1/3,0,0)$
-
-* $\mathbf{w}_2 = (1/2,1/2, 0, 0)$
-
-* $\mathbf{w}_3 = (1/3,1/3,1/3,0,0)$. 
-
-We will use a fixed correlation structure such that the correlation of chemicals within a group is .7 and the cross correlations between 2 chemicals in different groups is .5. 
-<br>
-We first define ESS and CCS. ESS is a list that contains the the sample size and the true underlying weight distribution of the chemicals in each group.  CCS is a two element vector where the first index defines the within group correlation and the second index controls the between group correlation.  The sample size for our dataset is fixed to $n = 10,000$ and we set the distribution of the true underlying weights as follows: 
-
-<br>
[…]
