+++ README.md
@@ -1,845 +1,0 @@
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-# fgwqsr – Frequentist Grouped Weighted Quantile Sum Regression
-
-Fully frequentist model fitting procedure for the weighted quantile sum
-regression formulation. Included in this package is an implementation of 
-both the Frequentist Grouped Weighted Quantile Sum Regression (FGWQSR) 
-and the Bayesian Grouped Weighted Quantile Sum Regression (bgwqsr) 
-introduced by Wheeler, David C et al. that fits Markov Chain Monte Carlo 
-(MCMC) chainsin parallel through the runjags package.
-
-<!-- badges: start -->
-<!-- badges: end -->
-
-# Overview
-
-<br> For a binary outcome, FGWQSR estimates parameters to the following model:
-
-<br>
-
-$$
-y_i \sim \text{Bernoulli}(\pi_i)
-$$
-
-$$
-\text{logit}(\pi_i) = c_0+ \sum\limits_{g = 1}^G \gamma_g \bigg( \sum\limits_{k = 1} ^{c_g} w_{g,k} \cdot q_{g,k,i}\bigg) + \sum_{r = 1}^R \phi_rz_{r,i} 
-$$
-
-where
-
-- $y_i$ - outcome variable (coded 0 1)
-
-- $c_0$ - offset variable (intercept)
-
-- $G$ - total number of chemical mixture groups
-
-- $\gamma_g$ - group index effect corresponding to mixture group $g$
-
-- $c_g$ - total number of chemicals in group $g$
-
-- $w_{g,k}$ - weight for chemical $k$ in group $g$
-
-- $q_{g,k,i}$ - quantized chemical exposure $k$ in group $g$ for
-  individual $i$
-
-- $R$ - total number of adjusting covariates
-
-- $\phi_r$ - effect of $r^{\text{th}}$ adjusting covariate
-
-- $z_{r,i}$ - measure of adjusting covariate $r$ for individual $i$
-
-subject to the constraints
-$$\sum_{k = 1}^{c_g} w_{g,k} =1, \ w_{g,k} \in (0,1)$$ without requiring
-any data splitting
-
-<br>
-
-# Installation
-
-You can install the development version of fgwqsr from
-[GitHub](https://github.com/) with:
-
-``` r
-# install.packages("devtools")
-devtools::install_github("Daniel-Rud/fgwqsr")
-#> Downloading GitHub repo Daniel-Rud/fgwqsr@HEAD
-#> fansi   (1.0.4 -> 1.0.5) [CRAN]
-#> withr   (2.5.0 -> 2.5.1) [CRAN]
-#> ggplot2 (3.4.3 -> 3.4.4) [CRAN]
-#> Installing 3 packages: fansi, withr, ggplot2
-#> Installing packages into '/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpVPL745/temp_libpath874955fed213'
-#> (as 'lib' is unspecified)
-#> 
-#> The downloaded binary packages are in
-#>  /var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T//RtmpP4mS6a/downloaded_packages
-#> ── R CMD build ─────────────────────────────────────────────────────────────────
-#> * checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpP4mS6a/remotes8ac9332aeec7/Daniel-Rud-fgwqsr-63eb157/DESCRIPTION’ ... OK
-#> * preparing ‘fgwqsr’:
-#> * checking DESCRIPTION meta-information ... OK
[…]
