+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# fgwqsr – Frequentist Grouped Weighted Quantile Sum Regression
+# fgwqsr – Frequentist Grouped Weighted Quantile Sum Regression <img src="man/figures/FGWQSR_hex.png" width="100" align="right" />
 
 Fully frequentist model fitting procedure for the weighted quantile sum
 regression formulation. Included in this package is an implementation of
@@ -21,6 +21,7 @@ in parallel through the runjags package.
 $$y_i \sim \text{Bernoulli}(\pi_i) $$
 
 $$\text{logit}(\pi_i) = c_0+ \sum\limits_{g = 1}^G \gamma_g \bigg( \sum\limits_{k = 1} ^{c_g} w_{g,k} \cdot q_{g,k,i}\bigg) + \sum_{r = 1}^R \phi_rz_{r,i} $$
+
 where
 
 - $y_i$ - outcome variable (coded 0 1)
@@ -62,9 +63,9 @@ devtools::install_github("Daniel-Rud/fgwqsr")
 #> Downloading GitHub repo Daniel-Rud/fgwqsr@HEAD
 #> 
 #> ── R CMD build ─────────────────────────────────────────────────────────────────
-#>      checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmprpMIRn/remotes16d8746461745/Daniel-Rud-fgwqsr-8f82fee/DESCRIPTION’ ...  ✔  checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmprpMIRn/remotes16d8746461745/Daniel-Rud-fgwqsr-8f82fee/DESCRIPTION’
+#>      checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpNaCE4k/remotes17c3f61554d60/Daniel-Rud-fgwqsr-52f0b8b/DESCRIPTION’ ...  ✔  checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpNaCE4k/remotes17c3f61554d60/Daniel-Rud-fgwqsr-52f0b8b/DESCRIPTION’
 #>   ─  preparing ‘fgwqsr’:
-#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
+#>    checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
 #>   ─  checking for LF line-endings in source and make files and shell scripts
 #>   ─  checking for empty or unneeded directories
 #>      Omitted ‘LazyData’ from DESCRIPTION
@@ -382,7 +383,7 @@ summary(fgwqsr_fit)
 #> 
 #> Significance Codes: <0.001 '***' <0.01 '**' <0.05 '*' <0.10 '.' 
 #> 
-#> Total runtime for FGWQSR:  1.27 minutes on 10 cores.
+#> Total runtime for FGWQSR:  1.67 minutes on 10 cores.
 ```
 
 <br> We can compare the true underlying group indices and chemical
@@ -554,7 +555,7 @@ summary(fgwqsr_fit_adj)
 #> 
 #> Significance Codes: <0.001 '***' <0.01 '**' <0.05 '*' <0.10 '.' 
 #> 
-#> Total runtime for FGWQSR:  1.35 minutes on 10 cores.
+#> Total runtime for FGWQSR:  1.77 minutes on 10 cores.
 ```
 
 <br>
[…]
