### 2025-05-12 60282c8
fixed poisson hessian function for memory efficiency.  Problem occured when constructing diag matrix for large number of observations.
