### 2025-05-17 #8012 @Kevin-Mattheus-Moerman
@Daniel-Rud I am the AEiC on this track and here to help with initial steps. It looks like you created a dedicated repository just for the JOSS paper work. This is not the correct approach. Instead you should add the JOSS related files to your main project repository (which I take it is this one: https://github.com/Daniel-Rud/fgwqsr ? ). You can either add it to the main branch or create a dedicated branch and let us know what branch you used. But please process this now.
