### 2025-05-21 #8012 @Daniel-Rud
@Kevin-Mattheus-Moerman  Thank you very much for the corrections.

1) I have added the joss paper to the  https://github.com/Daniel-Rud/fgwqsr repository, it is in the branch titled joss-branch.  Would you mind updating the project repository, I do not think I have the permission to make those changes.

2) The published manuscript can be found here: https://onlinelibrary.wiley.com/doi/full/10.1002/sim.70078
