### 2025-06-10 #8012 @goldingn
Hi @blind-contours, would you be interested in reviewing this R package submission on weighted quantile sum regression for JOSS?
