### 2025-07-27 #8012 @goldingn
Hi @jamesuanhoro would you be interested in reviewing this R package submission on weighted quantile sum regression for JOSS?
