### 2025-08-01 #8012 @goldingn
Wonderful, thank you @msalibian and @ibarraespinosa! I'll add you as reviewers now, and kick off the review process.

You should be added to a new GitHub issue where the review will take place. You'll just need to run an editorial-bot command to generate your review checklists and then you'll be able to start reviewing.
