### 2025-08-01 #8692 @editorialbot
[REVIEW]: fgwqsr: An R package for Frequentist Grouped Weighted Quantile Sum Regression
**Submitting author:** @Daniel-Rud (<a href="http://orcid.org/0000-0002-0508-4552">Daniel Rud</a>)
**Repository:**  https://github.com/Daniel-Rud/fgwqsr
**Branch with paper.md** (empty if default branch): joss-branch
**Version:** v1.0.1

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the <a href=" https://github.com/Daniel-Rud/fgwqsr" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
