### 2025-08-08 Daniel-Rud/fgwqsr#3 @ibarraespinosa
review for joss
hi I have started the review for joss

https://github.com/openjournals/joss-reviews/issues/8692
