### 2025-08-08 #8692 @msalibian
@Daniel-Rud The paper looks fine to me, but it's missing a "Community guidelines" section, which I believe is needed in all submissions. Perhaps @goldingn can confirm?
