### 2025-08-09 Daniel-Rud/fgwqsr#3 @ibarraespinosa
you could  change

```devtools::install_github("Daniel-Rud/fgwqsr")```

to

```remotes::install_github("Daniel-Rud/fgwqsr")```
