### 2025-08-09 Daniel-Rud/fgwqsr#3 @ibarraespinosa
Must in cldue in readme that depends on jags. Maybe also in DESCRIPTION
