### 2025-08-09 Daniel-Rud/fgwqsr#3 @ibarraespinosa
```
The downloaded source packages are in
	‘/tmp/RtmpiTBA1u/downloaded_packages’
── R CMD build ────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/tmp/RtmpiTBA1u/remotes260d2e0f6937/Daniel-Rud-fgwqsr-60282c8/DESCRIPTION’ ...
─  preparing ‘fgwqsr’:
   Omitted ‘LazyData’ from DESCRIPTION
─  building ‘fgwqsr_0.1.0.tar.gz’

Installing package into ‘/home/sergio/R/x86_64-pc-linux-gnu-library/4.4’
(as ‘lib’ is unspecified)
* installing *source* package ‘fgwqsr’ ...
** using staged installation
** R
** byte-compile and prepare package for lazy loading
** help
*** installing help indices
*** copying figures
** building package indices
** installing vignettes
** testing if installed package can be loaded from temporary location
** testing if installed package can be loaded from final location
** testing if installed package keeps a record of temporary installation path
* DONE (fgwqsr)
 ```

Check LazyData guidelines for DESCRIPTION
