### 2025-08-09 Daniel-Rud/fgwqsr#3 @ibarraespinosa
Since you have summary and print methods, why dont you add plot methods?

```
 plot(fgwqsr_fit)
Error in xy.coords(x, y, xlabel, ylabel, log) :
  'x' is a list, but does not have components 'x' and 'y'
```
