### 2025-08-09 Daniel-Rud/fgwqsr#3 @ibarraespinosa
The readme seems part of a manuscript under review elsewhere. You have city A and b, and pollutants like that. However, it seems that if you named the name of the city and the pollutants witgh they respective names, it would provide a clearer understanding of your methods. In general package looks ok, I would suggest some minor changes in the coding  style, also, plotting methods. That said, everything looks ok
