### 2025-08-11 #8692 @goldingn
Thanks for the rapid review @msalibian!

Just to clarify: the paper doesn't need that section, but the information should be present and findable in the repo. E.g. something in the README letting people know how they can contribute to the project along these lines: https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines
