### 2025-08-11 #8692 @msalibian
@Daniel-Rud please see the comment [above](https://github.com/openjournals/joss-reviews/issues/8692#issuecomment-3173025294), and either add a Community Guidelines to the paper, or else, include the relevant information in the README file, along the suggestions in https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines.
