### 2025-08-24 Daniel-Rud/fgwqsr#3 @Daniel-Rud
@ibarraespinosa Thank you so much, sorry for the delay, I will respond to these comments next weekend.
