+++ paper.md
@@ -101,8 +101,8 @@ Thus, FGWQSR represents a significant advancement in WQSR methodology, providing
 
 The most current version of `fgwqsr` package can be downloaded from github using the following instructions: 
 ```r
-install.packages("devtools")
-devtools::install_github("Daniel-Rud/fgwqsr")
+install.packages("remotes")
+remotes::install_github("Daniel-Rud/fgwqsr")
 ```
 
 # References
