### 2025-08-31 cd687ed
added plot function
