+++ README.Rmd
@@ -14,7 +14,9 @@ knitr::opts_chunk$set(
 ```
 
 # fgwqsr -- Frequentist Grouped Weighted Quantile Sum Regression <img src="man/figures/FGWQSR_hex.png" width="100" align="right" />
-Fully frequentist model fitting procedure for the weighted quantile sum regression formulation.  Included in this package is an implementation of Bayesian Grouped Weighted Quantile Sum Regression (bgwqsr) introduced by Wheeler, David C et al. that fits Markov Chain Monte Carlo (MCMC) chains in parallel through the runjags package.  
+Fully frequentist estimation and inference for Weighted Quantile Sum Regression, enabling inference on both individual and grouped pollutants without the need for data splitting. [Rud et al (2025)](https://onlinelibrary.wiley.com/doi/full/10.1002/sim.70078).  
+
+Also included in this package is an implementation of Bayesian Grouped Weighted Quantile Sum Regression (bgwqsr) introduced by Wheeler, David C et al. that fits Markov Chain Monte Carlo (MCMC) chains in parallel through the runjags package.  
 
 <!-- badges: start -->
 <!-- badges: end -->
@@ -62,16 +64,18 @@ subject to the constraints $$\sum_{k = 1}^{c_g} w_{g,k} =1, \ w_{g,k} \in (0,1)$
 
 # Installation
 
-You can install the development version of fgwqsr from [GitHub](https://github.com/) with:
+You can install the development version of fgwqsr from [GitHub](https://github.com/Daniel-Rud/fgwqsr) with:
 
 ```{r}
 # install.packages("devtools")
-devtools::install_github("Daniel-Rud/fgwqsr")
+remotes::install_github("Daniel-Rud/fgwqsr")
 ```
 
 ```{r example}
 library(fgwqsr)
 ```
+
+The `fgwqsr` package depends on JAGS installation through the `runjags` package, which is used to fit BGWQSR models.  Please see the `runjags` installation instructions for your operating system [here](https://cran.r-project.org/web/packages/runjags/vignettes/Installation.html).  Note that JAGS is not required to fit FGWQSR models.
 
 In order to use fgwqsr, let us first generate some data
 
@@ -100,8 +104,8 @@ ess = list(n = n,
            weights = list(w1 = c(1/3,1/3,1/3,0,0), 
                           w2 = c(1/2,1/2, 0, 0), 
                           w3 = c(1/3,1/3,1/3,0,0)
-                          )
            )
+)
 
 ccs = c(.5,.1)
 ```
@@ -159,10 +163,26 @@ chem_data = mvtnorm::rmvnorm(n, mean = rep(0, 14), sigma = corr)
 ```
 <br>
 
-Next, we quantize the exposure variables in order to make comparisons between the true underlying oracle model and FGWQSR estimates.  Note that we will subtract 1 from our quantized variables, so that the first quintile corresponds to the measure of 0.  By doing this, we can ensure that we have an interpretable model intercept.  
+Next, we quantize the exposure variables in order to make comparisons between the true underlying oracle model and FGWQSR estimates.  Note that we will subtract 1 from our quantized variables, so that the first quintile corresponds to the measure of 0.  By doing this, we can ensure that we have an interpretable model intercept. For ease of understanding, we give the chemical variables arbitrary names of chemical constituents.   
 
 ```{r}
 chem_data = apply(chem_data, MARGIN = 2, statar::xtile, n = 5) - 1
+colnames(chem_data) = chemicals <- c(
+  "PM2.5", 
+  "NO2", 
+  "O3", 
+  "SO2", 
+  "CO", 
+  "Lead", 
+  "Mercury", 
+  "Arsenic", 
+  "Cadmium", 
+  "Benzene", 
+  "Toluene", 
+  "Phthalates", 
[…]
