+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(plot,fgwqsr)
 S3method(print,fgwqsr)
 S3method(summary,fgwqsr)
 export("%>%")
