### 2025-08-31 8814766
updated README with plot functions and changes to pollutant names and city names
