+++ README.md
@@ -3,11 +3,15 @@
 
 # fgwqsr – Frequentist Grouped Weighted Quantile Sum Regression <img src="man/figures/FGWQSR_hex.png" width="100" align="right" />
 
-Fully frequentist model fitting procedure for the weighted quantile sum
-regression formulation. Included in this package is an implementation of
-Bayesian Grouped Weighted Quantile Sum Regression (bgwqsr) introduced by
-Wheeler, David C et al. that fits Markov Chain Monte Carlo (MCMC) chains
-in parallel through the runjags package.
+Fully frequentist estimation and inference for Weighted Quantile Sum
+Regression, enabling inference on both individual and grouped pollutants
+without the need for data splitting. [Rud et al
+(2025)](https://onlinelibrary.wiley.com/doi/full/10.1002/sim.70078).
+
+Also included in this package is an implementation of Bayesian Grouped
+Weighted Quantile Sum Regression (bgwqsr) introduced by Wheeler, David C
+et al. that fits Markov Chain Monte Carlo (MCMC) chains in parallel
+through the runjags package.
 
 <!-- badges: start -->
 <!-- badges: end -->
@@ -54,29 +58,25 @@ any data splitting
 # Installation
 
 You can install the development version of fgwqsr from
-[GitHub](https://github.com/) with:
+[GitHub](https://github.com/Daniel-Rud/fgwqsr) with:
 
 ``` r
 # install.packages("devtools")
-devtools::install_github("Daniel-Rud/fgwqsr")
+remotes::install_github("Daniel-Rud/fgwqsr")
 #> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
-#> Downloading GitHub repo Daniel-Rud/fgwqsr@HEAD
-#> 
-#> ── R CMD build ─────────────────────────────────────────────────────────────────
-#>      checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpNaCE4k/remotes17c3f61554d60/Daniel-Rud-fgwqsr-52f0b8b/DESCRIPTION’ ...  ✔  checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpNaCE4k/remotes17c3f61554d60/Daniel-Rud-fgwqsr-52f0b8b/DESCRIPTION’
-#>   ─  preparing ‘fgwqsr’:
-#>    checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
-#>   ─  checking for LF line-endings in source and make files and shell scripts
-#>   ─  checking for empty or unneeded directories
-#>      Omitted ‘LazyData’ from DESCRIPTION
-#>   ─  building ‘fgwqsr_0.1.0.tar.gz’
-#>      
-#> 
+#> Skipping install of 'fgwqsr' from a github remote, the SHA1 (883d80c2) has not changed since last install.
+#>   Use `force = TRUE` to force installation
 ```
 
 ``` r
 library(fgwqsr)
 ```
+
+The `fgwqsr` package depends on JAGS installation through the `runjags`
+package, which is used to fit BGWQSR models. Please see the `runjags`
+installation instructions for your operating system
+[here](https://cran.r-project.org/web/packages/runjags/vignettes/Installation.html).
+Note that JAGS is not required to fit FGWQSR models.
 
 In order to use fgwqsr, let us first generate some data
 
@@ -115,8 +115,8 @@ ess = list(n = n,
[…]
