### 2025-08-31 #8692 @Daniel-Rud
Hi @msalibian, I have added a community guidelines section the README file in the main branch of fgwqsr:
https://github.com/Daniel-Rud/fgwqsr
