### 2025-08-31 d264efa
set LazyData setting to false
