+++ DESCRIPTION
@@ -43,5 +43,5 @@ VignetteBuilder:
     runjags
 Config/testthat/edition: 3
 Encoding: UTF-8
-LazyData: true
+LazyData: false
 RoxygenNote: 7.3.2
