+++ README.md
@@ -64,8 +64,18 @@ You can install the development version of fgwqsr from
 # install.packages("devtools")
 remotes::install_github("Daniel-Rud/fgwqsr")
 #> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
-#> Skipping install of 'fgwqsr' from a github remote, the SHA1 (883d80c2) has not changed since last install.
-#>   Use `force = TRUE` to force installation
+#> Downloading GitHub repo Daniel-Rud/fgwqsr@HEAD
+#> 
+#> ── R CMD build ─────────────────────────────────────────────────────────────────
+#>      checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpelvZTp/remotes121c86043a123/Daniel-Rud-fgwqsr-8814766/DESCRIPTION’ ...  ✔  checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpelvZTp/remotes121c86043a123/Daniel-Rud-fgwqsr-8814766/DESCRIPTION’
+#>   ─  preparing ‘fgwqsr’:
+#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
+#>   ─  checking for LF line-endings in source and make files and shell scripts
+#>   ─  checking for empty or unneeded directories
+#>      Omitted ‘LazyData’ from DESCRIPTION
+#>   ─  building ‘fgwqsr_0.1.0.tar.gz’
+#>      
+#> 
 ```
 
 ``` r
@@ -415,7 +425,7 @@ summary(fgwqsr_fit)
 #> 
 #> Significance Codes: <0.001 '***' <0.01 '**' <0.05 '*' <0.10 '.' 
 #> 
-#> Total runtime for FGWQSR:  1.41 minutes on 10 cores.
+#> Total runtime for FGWQSR:  1.42 minutes on 10 cores.
 ```
 
 <br>
@@ -604,7 +614,7 @@ summary(fgwqsr_fit_adj)
 #> 
 #> Significance Codes: <0.001 '***' <0.01 '**' <0.05 '*' <0.10 '.' 
 #> 
-#> Total runtime for FGWQSR:  1.54 minutes on 10 cores.
+#> Total runtime for FGWQSR:  1.57 minutes on 10 cores.
 ```
 
 <br>
@@ -733,7 +743,7 @@ bgwqsr_fit = bgwqsr(formula = mod_formula_adj,
 #> Calling 3 simulations using the parallel method...
 #> Following the progress of chain 1 (the program will wait for all chains
 #> to finish before continuing):
-#> Welcome to JAGS 4.3.2 (official binary) on Sun Aug 31 12:02:50 2025
+#> Welcome to JAGS 4.3.2 (official binary) on Sun Aug 31 12:13:59 2025
 #> JAGS is free software and comes with ABSOLUTELY NO WARRANTY
 #> Loading module: basemod: ok
 #> Loading module: bugs: ok
@@ -777,72 +787,72 @@ the summaries object from the bgwqsr model.
 
 ``` r
 bgwqsr_fit$model$summaries
-#>                              Lower95        Median    Upper95         Mean
-#> B0                      -4.43934e+00 -3.2593150000 -2.0041900 -3.249777458
-#> B1                      -7.56010e-01 -0.7136870000 -0.6705880 -0.713888593
-#> B2                      -4.61841e-02 -0.0051817650  0.0328981 -0.005899786
[…]
