### 2025-08-31 Daniel-Rud/fgwqsr#3 @Daniel-Rud
Hi @ibarraespinosa ,
I have added the following changes:

1) Added a plot function for fgwqsr.  Please see the changes in the README file and I have added the call in the joss paper.
2) I have changed the city names and the pollutant names to resemble realistic analysis.
3) I have included in the README Installation section that the package depends on JAGS installation
4) I set the LazyData setting in the DESCRIPTION file to false.
