### 2025-09-04 Daniel-Rud/fgwqsr#3 @ibarraespinosa
1. The package feels a little minimal, you have main functions and a few methods, but it would be a pain using this routinely because there is a lot here that needs very exact inputs.
2. Although authors have included examples in the readme and vignettes, they could use small examples for each on the main function (e.g. under ?fgwqsr and ?bgwqsr). Authors are proposing submitting to CRAN and think they may also need to add quick examples on-borad documentation to pass a CRAN review as well..
3. readme was quite involved. It felt like part of a larger example that needed more explanation or scaling down to a quicker/shorter example. Probably the latter better idea and turning current readme into a vignette?
4. Authors should add missing DOIs (first four paper in references) for consistency.
