### 2025-09-08 7cd5742
added doi's for references with missing doi's.
