+++ paper.bib
@@ -23,6 +23,7 @@ year = {2025}
   number={1},
   pages={100--120},
   year={2015},
+  doi = {https://doi.org/10.1007/s13253-014-0180-3}
 }
 
 
@@ -72,8 +73,9 @@ JOURNAL = {Current epidemiology reports},
 PUBLISHER = {Springer},
 VOLUME = {5},
 YEAR = {2018},
-NUMBER = {2}, 
-PAGES = {160--165}}
+NUMBER = {2},
+PAGES = {160--165},
+DOI = {https://doi.org/10.1007/s40471-018-0145-0}}
 
 
 @Article{an_overview_of_methods,
@@ -83,7 +85,8 @@ JOURNAL = {Environmental Health},
 PUBLISHER = {BMC},
 VOLUME = {18},
 YEAR = {2019},
-NUMBER = {76}}
+NUMBER = {76},
+DOI = {https://doi.org/10.1186/s12940-019-0515-1}}
 
 @Article{using_three_statistical,
 AUTHOR = {Bangsheng Wu and Yi Jiang and Xiaoqing Jin and Li He},
@@ -92,7 +95,8 @@ JOURNAL = {Environmental Health},
 PUBLISHER = {BMC},
 VOLUME = {19},
 YEAR = {2020},
-NUMBER = {1}}
+NUMBER = {1},
+DOI = {https://doi.org/10.1186/s12940-020-00642-6}}
 
 @Article{air_pol_diabetes,
 AUTHOR = {Yongze Li and Lu Xu and Zhongyan Shan and Weiping Teng and Cheng Han},
@@ -101,7 +105,8 @@ JOURNAL = {Therapeutic Advances in Endocrinology and Metabolism},
 PUBLISHER = {SAGE. Publications},
 VOLUME = {10},
 YEAR = {2019},
-PAGES = {1--15}}
+PAGES = {1--15},
+DOI = {https://doi.org/10.1177/2042018819897046}}
 
 @Article{air_pol_lymphoma,
 AUTHOR = {Jenna Czarnota and Chris Gennings and Joanne S. Colt and Anneclaire J. De Roos and James R. Cerhan and
@@ -112,8 +117,9 @@ JOURNAL = {Environmental Health Perspectives},
 PUBLISHER = {NIEHS},
 VOLUME = {123},
 YEAR = {2015},
-NUMBER = {10}, 
-PAGES = {965--970}}
+NUMBER = {10},
+PAGES = {965--970},
+DOI = {https://doi.org/10.1289/ehp.1408630}}
 
 @Article{air_pol_obesity,
 AUTHOR = {Jeroen de Bont and  Yesika Díaz and Montserrat de Castro and Marta Cirach and Xavier Basagaña and Mark Nieuwenhuijsen and Talita Duarte-Salles and Martine Vrijheid},
@@ -122,7 +128,8 @@ JOURNAL = {International Journal of Obesity},
 PUBLISHER = {Nature},
 VOLUME = {45},
 YEAR = {2021},
-PAGES = {1124--1132}}
+PAGES = {1124--1132},
+DOI = {https://doi.org/10.1038/s41366-021-00783-9}}
 
 @Article{human_chemical_effect,
 AUTHOR = {David O. Carpenter and Kathleen Arcaro and David C. Spink},
@@ -131,7 +138,8 @@ JOURNAL = {Environmental Health Perspectives},
 PUBLISHER = {NIEHS},
 VOLUME = {110},
 YEAR = {2002},
-PAGES = {25--42}}
+PAGES = {25--42},
+DOI = {https://doi.org/10.1289/ehp.02110s125}}
 
 @Article{human_chemical_exposure,
 AUTHOR = {Ilona Silins and Johan Högberg},
[…]
