### 2025-09-08 Daniel-Rud/fgwqsr#3 @Daniel-Rud
Hi @ibarraespinosa ,

Thank you for the suggestions.  Responding to each point:

1. Since weighted quantile sum regression is already a well-established area, our goal with this package was to focus on our implementation of parameter estimation / inference for weighted quantile sum regression rather than building a large framework around it. The inputs are necessarily specific because they reflect the structure required for WQS, and we wanted to keep the package lightweight while documenting the workflow so that users can apply it confidently.
2.  I agree that there is value in having quick examples right in the function documentation. For now, we’ve kept the examples in the README and vignettes so users can see the full workflow in one place without repeating the same code in multiple spots. If it turns out CRAN really does require shorter inline examples, or if users let us know they’d find it helpful, we can definitely add them later.
3. The README and vignette intentionally contain the same material. Because of the nature of the subject area, illustrating use of the package requires somewhat involved examples to reflect realistic workflows and inputs. We felt this level of detail was important for demonstrating practical use, rather than relying on overly simplified toy cases.
4. I have added the missing DOIs for each reference that was previously missing a DOI.
