### 2025-09-21 #8692 @ibarraespinosa
paper looks good for publication. I suggest acceptance
