### 2025-09-24 #8692 @goldingn
Hi @Daniel-Rud please could you action the tasks above, so I can progress the acceptance process? Note the need for the archive to have the same authors as the paper, and the same license as the software.

I've opened an issue (link above) for one minor edit you could make before archiving. Otherwise I think the paper text looks great!
