### 2025-09-27 3f4d122
Changed instances of "T" to "TRUE"
