### 2025-09-27 #8692 @Daniel-Rud
Hi @goldingn,
I have changed all instances of "T" to "TRUE". Thank you for this explanation and bringing up this concern.
