### 2025-09-30 #8692 @goldingn
@Daniel-Rud The MIT license in the GitHub release is fine. You just need to update the Zenodo archive to have the same license.
I can see there (in the bar on the right of the zenodo page https://zenodo.org/records/17215639) that the archive is creative commons licensed, so there is currently a mismatch.
