### 2025-10-25 2d355f2
Changed instances of T to TRUE
