+++ README.md
@@ -65,9 +65,20 @@ You can install the development version of fgwqsr from
 remotes::install_github("Daniel-Rud/fgwqsr")
 #> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
 #> Downloading GitHub repo Daniel-Rud/fgwqsr@HEAD
+#> magrittr  (2.0.3  -> 2.0.4 ) [CRAN]
+#> S7        (NA     -> 0.2.0 ) [CRAN]
+#> pillar    (1.11.0 -> 1.11.1) [CRAN]
+#> hms       (1.1.3  -> 1.1.4 ) [CRAN]
+#> stringr   (1.5.1  -> 1.5.2 ) [CRAN]
+#> ggplot2   (3.5.2  -> 4.0.0 ) [CRAN]
+#> progressr (0.15.1 -> 0.17.0) [CRAN]
+#> pracma    (2.4.4  -> 2.4.6 ) [CRAN]
+#> Installing 8 packages: magrittr, S7, pillar, hms, stringr, ggplot2, progressr, pracma
 #> 
+#> The downloaded binary packages are in
+#>  /var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T//Rtmp7Cycnr/downloaded_packages
 #> ── R CMD build ─────────────────────────────────────────────────────────────────
-#>      checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpelvZTp/remotes121c86043a123/Daniel-Rud-fgwqsr-8814766/DESCRIPTION’ ...  ✔  checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/RtmpelvZTp/remotes121c86043a123/Daniel-Rud-fgwqsr-8814766/DESCRIPTION’
+#>      checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/Rtmp7Cycnr/remotes14bb165c2bf09/Daniel-Rud-fgwqsr-3f4d122/DESCRIPTION’ ...  ✔  checking for file ‘/private/var/folders/b4/9kg7p6cj729_pzc5dggk_9jm0000gn/T/Rtmp7Cycnr/remotes14bb165c2bf09/Daniel-Rud-fgwqsr-3f4d122/DESCRIPTION’
 #>   ─  preparing ‘fgwqsr’:
 #>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
 #>   ─  checking for LF line-endings in source and make files and shell scripts
@@ -367,7 +378,7 @@ fgwqsr_fit = fgwqsr(formula = mod_formula,
                     quantiles = 5,
                     family = 'binomial',
                     n_mvn_sims = 10000,
-                    verbose = T)
+                    verbose = TRUE)
 #> Fitting full and nested FGWQSR models...
 #> 
 #> Performing Likelihood Ratio Test Inference...
@@ -425,7 +436,7 @@ summary(fgwqsr_fit)
 #> 
 #> Significance Codes: <0.001 '***' <0.01 '**' <0.05 '*' <0.10 '.' 
 #> 
-#> Total runtime for FGWQSR:  1.42 minutes on 10 cores.
+#> Total runtime for FGWQSR:  1.54 minutes on 10 cores.
 ```
 
 <br>
@@ -501,7 +512,7 @@ the comparison between Santa Barbara and Los Angeles.
 set.seed(11)
 # create adjusting covariates
 weight = rnorm(n = n, mean = 68, sd = 2.5)
[…]
