### 2025-10-25 a5dd3f3
Removed unused references, changed DOIs to proper format
