+++ paper.bib
@@ -6,7 +6,7 @@
   number={1},
   pages={100--120},
   year={2015},
-  doi = {https://doi.org/10.1007/s13253-014-0180-3}
+  doi = {10.1007/s13253-014-0180-3}
 }
 
 @article{fgwqsr,
