### 2025-10-25 ff85bae
Changed one instance of T -> TRUE
